1582 Break circular dependency daoSerializer-ClassificatorServiceImpl + remove extra blank lines
