+++ classificators/build.gradle
@@ -32,7 +32,6 @@ dependencies {
     testCompile "org.springframework.restdocs:spring-restdocs-mockmvc:${project.ext['springRestdocsVersion']}"
 }
 
-
 test {
     outputs.dir snippetsDir
 }
