+++ classificators/src/main/java/ru/okpdmarket/controller/ClassificatorController.java
@@ -1,7 +1,12 @@
 package ru.okpdmarket.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 import ru.okpdmarket.model.Classificator;
 import ru.okpdmarket.model.ClassificatorItem;
 import ru.okpdmarket.service.ClassificatorService;
@@ -30,7 +35,6 @@ public class ClassificatorController {
     public List<Classificator> getClassificatorTypes() {
         return classificatorService.getClassificatorTypes();
     }
-
 
     @RequestMapping(value = "/{classificatorCode}", method = RequestMethod.GET)
     public List<ClassificatorItem> getTopItems(@PathVariable(value = "classificatorCode") String classificatorCode) {
