+++ classificators/src/main/java/ru/okpdmarket/dao/DaoSerializer.java
@@ -1,15 +1,12 @@
 package ru.okpdmarket.dao;
 
 import lombok.val;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.okpdmarket.dao.dto.ClassificatorDaoDto;
 import ru.okpdmarket.dao.dto.ClassificatorItemDaoDto;
 import ru.okpdmarket.dao.dto.ClassificatorLinkDaoDto;
 import ru.okpdmarket.model.Classificator;
 import ru.okpdmarket.model.ClassificatorItem;
-import ru.okpdmarket.service.ClassificatorService;
-import ru.okpdmarket.service.impl.ClassificatorItemService;
 
 import java.util.Collection;
 import java.util.List;
@@ -22,13 +19,19 @@ import java.util.stream.Stream;
 @Service
 public class DaoSerializer {
 
-    private final ClassificatorService service;
-    private final ClassificatorItemService itemService;
+    public List<ClassificatorDaoDto> serializeList(List<Classificator> items) {
+        return items.stream().map(this::serialize).collect(Collectors.toList());
+    }
 
-    @Autowired
-    public DaoSerializer(ClassificatorService service, ClassificatorItemService itemService) {
-        this.service = service;
-        this.itemService = itemService;
+    private ClassificatorDaoDto serialize(Classificator item) {
+        ClassificatorDaoDto dto = new ClassificatorDaoDto();
+        dto.setCode(item.getCode());
+        dto.setName(item.getName());
+        dto.setDescription(item.getDescription());
+        val daoDtoChilderList = serializeList(item.getContents().getFirstLevel());
+        dto.setTree(daoDtoChilderList);
+        dto.setLinks(serializeLinks(item));
+        return dto;
     }
 
     private ClassificatorItemDaoDto serialize(ClassificatorItem item) {
@@ -44,48 +47,6 @@ public class DaoSerializer {
     public List<ClassificatorItemDaoDto> serializeList(Collection<ClassificatorItem> items) {
         return items.stream().map(this::serialize).collect(Collectors.toList());
     }
-
-    private ClassificatorDaoDto serialize(Classificator item) {
-        ClassificatorDaoDto dto = new ClassificatorDaoDto();
-        dto.setCode(item.getCode());
-        dto.setName(item.getName());
-        dto.setDescription(item.getDescription());
-        val daoDtoChilderList = serializeList(item.getContents().getFirstLevel());
-        dto.setTree(daoDtoChilderList);
-        dto.setLinks(serializeLinks(item));
-        return dto;
-    }
-
-    public List<ClassificatorDaoDto> serializeList(List<Classificator> items) {
-        return items.stream().map(this::serialize).collect(Collectors.toList());
-    }
-
-    public Classificator deserialize(ClassificatorDaoDto dto) {
-        Classificator classificator = new Classificator();
-        classificator.setCode(dto.getCode());
-        classificator.setName(dto.getName());
-        classificator.setDescription(dto.getDescription());
-        service.put(classificator);
-        loadChildren(null, dto.getTree(), classificator);
-        return classificator;
-    }
-
-    private void loadChildren(String parentCode, List<ClassificatorItemDaoDto> fromDtoList, Classificator targetClassificator) {
-        for (ClassificatorItemDaoDto childDto : fromDtoList) {
-            val item = new ClassificatorItem(childDto.getCode(), childDto.getName(), childDto.getNotes());
-            item.setParentCode(parentCode);
-            itemService.addItem(targetClassificator.getCode(), item);
-
-            if (!childDto.getChildren().isEmpty()) {
-                loadChildren(childDto.getCode(), childDto.getChildren(), targetClassificator);
-            }
-        }
-    }
-
-    public List<Classificator> deserializeList(List<ClassificatorDaoDto> dtos) {
-        return dtos.stream().map(this::deserialize).collect(Collectors.toList());
-    }
-
 
     public List<ClassificatorLinkDaoDto> serializeLinks(Classificator classificator) {
         return classificator.getContents().getElements().values().stream()
@@ -107,15 +68,4 @@ public class DaoSerializer {
         return src.getCode().compareTo(dst.getCode()) > 0;
     }
 
-    public void loadLinks(List<ClassificatorLinkDaoDto> links) {
-        if (links != null)
-            links.forEach(this::loadLink);
-
-    }
-
-    private void loadLink(ClassificatorLinkDaoDto link) {
-        val sourceItem = service.getItem(link.getSrcClsCode(), link.getSrcItemCode());
-        val targetItem = service.getItem(link.getDstClsCode(), link.getDstItemCode());
-        itemService.linkItem(sourceItem, targetItem);
-    }
 }
