+++ classificators/src/main/java/ru/okpdmarket/model/Classificator.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 import ru.okpdmarket.repository.impl.ClassificatorContents;
 
-
 /**
  * Classificator domain model
  * Created by Vladislav on 29.08.2016.
