+++ classificators/src/main/java/ru/okpdmarket/model/ClassificatorItemRelations.java
@@ -33,5 +33,4 @@ public class ClassificatorItemRelations {
     }*/
 
 
-
 }
