+++ classificators/src/main/java/ru/okpdmarket/model/ClassificatorLinks.java
@@ -7,5 +7,4 @@ import java.util.concurrent.CopyOnWriteArrayList;
  */
 public class ClassificatorLinks extends CopyOnWriteArrayList<ClassificatorItem> {
 
-
 }
