+++ classificators/src/main/java/ru/okpdmarket/repository/ClassificatorRepository.java
@@ -4,7 +4,6 @@ import ru.okpdmarket.model.Classificator;
 
 import java.util.List;
 
-
 public interface ClassificatorRepository {
 
     List<Classificator> getClassificators();
