+++ classificators/src/main/java/ru/okpdmarket/service/SearchService.java
@@ -15,5 +15,4 @@ public interface SearchService {
 
     void indexClassificator(Classificator classificator);
 
-
 }
