+++ classificators/src/test/java/ru/okpdmarket/OkpdMarketApplicationTests.java
@@ -2,7 +2,6 @@ package ru.okpdmarket;
 
 import org.junit.Test;
 
-
 public class OkpdMarketApplicationTests extends IntegrationTest {
 
    /* @Configuration
