+++ classificators/src/test/java/ru/okpdmarket/dao/ClassificatorDaoTest.java
@@ -1,6 +1,5 @@
 package ru.okpdmarket.dao;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import ru.okpdmarket.IntegrationTest;
@@ -18,9 +17,7 @@ public class ClassificatorDaoTest extends IntegrationTest {
     @Autowired
     ClassificatorDao classificatorDao;
 
-
     @Test
-    @Ignore
     public void addClassificator(){
         ClassificatorDaoDto classificator = new ClassificatorDaoDto();
         classificator.setCode("code");
@@ -72,6 +69,5 @@ public class ClassificatorDaoTest extends IntegrationTest {
         //Classificator classificator1 = classificatorDao.findOne("test");
         //assertEquals(classificator.getCode(), classificator1.getCode());
     }
-
 
 }
