+++ config/build.gradle
@@ -28,4 +28,3 @@ dependencies {
 }
 
 
-
